requirements.txt
aiogram==3.*
python-dotenv
APScheduler
pytz
python-telegram-bot==20.0b1